OscCommandService > ExecuteOscCommands

+++ Hoscy/Services/Osc/Misc/OscCommandInfo.cs
@@ -7,7 +7,7 @@ public record OscCommandInfo()
     public required string Address;
     public required object[] Arguments;
     public string? Ip;
-    public int? Port;
+    public ushort? Port;
     public int? Wait;
 
     override public string ToString()
